ISSUE #682 Moved some utils into common resource and built out unit tests for mesh map reorganiser. Fixed bug in usages array.
