+++ test/src/unit/CMakeLists.txt
@@ -5,10 +5,16 @@
 
 
 add_subdirectory(repo)
+set(TEST_SOURCES
+	${TEST_SOURCES}
+	${CMAKE_CURRENT_SOURCE_DIR}/repo_test_mesh_utils.cpp
+	CACHE STRING "TEST_SOURCES" FORCE)
+
 set(TEST_HEADERS
 	${TEST_HEADERS}
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_test_database_info.h
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_test_fileservice_info.h
+	${CMAKE_CURRENT_SOURCE_DIR}/repo_test_mesh_utils.h
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_test_utils.h
 	CACHE STRING "TEST_HEADERS" FORCE)
