+++ test/src/unit/repo/manipulator/modeloptimizer/ut_repo_optimizer_multipart.cpp
@@ -18,11 +18,11 @@
 #include <gtest/gtest.h>
 #include <cstdlib>
 #include <repo/manipulator/modeloptimizer/repo_optimizer_multipart.h>
-#include <repo/core/model/bson/repo_bson_factory.h>
 #include <limits>
-#include <boost/functional/hash.hpp>
+#include <test/src/unit/repo_test_mesh_utils.h>
 
 using namespace repo::manipulator::modeloptimizer;
+using namespace repo::test::utils::mesh;
 
 const static repo::core::model::RepoScene::GraphType DEFAULT_GRAPH = repo::core::model::RepoScene::GraphType::DEFAULT;
 const static repo::core::model::RepoScene::GraphType OPTIMIZED_GRAPH = repo::core::model::RepoScene::GraphType::OPTIMIZED;
@@ -50,208 +50,9 @@ TEST(MultipartOptimizer, ApplyOptimizationTestEmpty)
 	delete empty2;
 }
 
-// Builds a triangle soup MeshNode with exactly nVertices, and faces
-// constructed such that each vertex is referenced at least once.
-
-repo::core::model::MeshNode* createRandomMesh(const int nVertices, const bool hasUV, const int primitiveSize, const std::vector<repo::lib::RepoUUID> &parent) {
-	std::vector<repo::lib::RepoVector3D> vertices;
-	std::vector<repo_face_t> faces;
-
-	std::vector<std::vector<float>> bbox = { 
-		{FLT_MAX, FLT_MAX, FLT_MAX},
-		{-FLT_MAX, -FLT_MAX, -FLT_MAX}
-	};
-
-	for (int i = 0; i < nVertices; ++i) {
-		repo::lib::RepoVector3D vertex = {	
-			static_cast<float>(std::rand()),
-			static_cast<float>(std::rand()),
-			static_cast<float>(std::rand()) 
-		};
-		vertices.push_back(vertex);
-
-		if (vertex.x < bbox[0][0])
-		{
-			bbox[0][0] = vertex.x;
-		}
-		if (vertex.y < bbox[0][1])
-		{
-			bbox[0][1] = vertex.y;
-		}
-		if (vertex.z < bbox[0][2])
-		{
-			bbox[0][2] = vertex.z;
-		}
-
-		if (vertex.x > bbox[1][0])
-		{
-			bbox[1][0] = vertex.x;
-		}
-		if (vertex.y > bbox[1][1])
-		{
-			bbox[1][1] = vertex.y;
-		}
-		if (vertex.z > bbox[1][2])
-		{
-			bbox[1][2] = vertex.z;
-		}
-	}
-
-	int vertexIndex = 0;
-	do
-	{
-		repo_face_t face;
-		for (int j = 0; j < primitiveSize; j++)
-		{
-			face.push_back((vertexIndex++ % vertices.size()));
-		}
-		faces.push_back(face);
-	} while (vertexIndex < vertices.size());
-
-	std::vector<std::vector<repo::lib::RepoVector2D>> uvs;
-	if (hasUV) {
-		std::vector<repo::lib::RepoVector2D> channel;
-		for (int i = 0; i < nVertices; ++i) {
-			channel.push_back({ (float)std::rand() / RAND_MAX, (float)std::rand() / RAND_MAX });
-		}
-		uvs.push_back(channel);
-	}
-
-	auto mesh = new repo::core::model::MeshNode(repo::core::model::RepoBSONFactory::makeMeshNode(vertices, faces, {}, bbox, uvs, {}, "mesh", parent));
-
-	return mesh;
-}
-
 // The functions below compare the geometry of the original MeshNodes with the 
 // stash MeshNodes as a triangle soup. The geometry should be identical.
 
-struct GenericFace
-{
-	std::vector<float> data;
-	int hit;
-
-	void push(repo::lib::RepoVector2D v)
-	{
-		data.push_back(v.x);
-		data.push_back(v.y);
-	}
-
-	void push(repo::lib::RepoVector3D v)
-	{
-		data.push_back(v.x);
-		data.push_back(v.y);
-		data.push_back(v.z);
-	}
-
-	const long hash() const
-	{
-		size_t hash = 0;
-		for (const auto value : data)
-		{
-			boost::hash_combine(hash, value);
-		}
-		return hash;
-	}
-
-	const bool equals(const GenericFace& other) const
-	{
-		return data == other.data;
-	}
-};
-
-void addFaces(repo::core::model::MeshNode* mesh, std::vector<GenericFace>& faces)
-{
-	if (mesh == nullptr)
-	{
-		return;
-	}
-
-	auto vertices = mesh->getVertices();
-	auto channels = mesh->getUVChannelsSeparated();
-	auto normals = mesh->getNormals();
-
-	for (const auto face : mesh->getFaces())
-	{
-		GenericFace portableFace;
-		portableFace.hit = 0;
-
-		for (const auto index : face)
-		{
-			portableFace.push(vertices[index]);
-
-			for (const auto channel : channels)
-			{
-				portableFace.push(channel[index]);
-			}
-
-			if (normals.size())
-			{
-				portableFace.push(normals[index]);
-			}
-		}
-
-		faces.push_back(portableFace);
-	}
-}
-
-
-bool compareMeshes(repo::core::model::RepoNodeSet original, repo::core::model::RepoNodeSet stash)
-{
-	std::vector<GenericFace> defaultFaces;
-	for (const auto node : original)
-	{		
-		addFaces(dynamic_cast<repo::core::model::MeshNode*>(node), defaultFaces);
-	}
-
-	std::vector<GenericFace> stashFaces;
-	for (const auto node : stash)
-	{
-		addFaces(dynamic_cast<repo::core::model::MeshNode*>(node), stashFaces);
-	}
-
-	if (defaultFaces.size() != stashFaces.size())
-	{
-		return false;
-	}
-
-	// Faces are compared exactly using a hash table for speed.
-
-	std::map<long, std::vector<GenericFace>> actual;
-
-	for (auto& face : stashFaces)
-	{
-		actual[face.hash()].push_back(face);
-	}
-
-	for (auto& face : defaultFaces)
-	{
-		auto& others = actual[face.hash()];
-		for (auto& other : others) {
-			if (other.hit) 
-			{
-				continue; // Each actual face may only be matched once
-			}
-			if (other.equals(face))
-			{
-				face.hit++;
-				other.hit++;
-				break;
-			}
-		}
-	}
-
-	// Did we find a match for all faces?
-
-	for (const auto& face : defaultFaces)
-	{
-		if (!face.hit)
-		{
-			return false;
-		}
-	}
-
-	return true;
-}
 
 TEST(MultipartOptimizer, TestAllMerged)
 {
