+++ test/src/unit/repo/manipulator/modelutility/CMakeLists.txt
@@ -4,6 +4,8 @@
 #If you really need to overwrite this file, be aware that it will be overwritten if updateSources.py is executed.
 
 
-add_subdirectory(modelconvertor)
-add_subdirectory(modeloptimizer)
-add_subdirectory(modelutility)
+set(TEST_SOURCES
+	${TEST_SOURCES}
+	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_mesh_map_reorganiser.cpp
+	CACHE STRING "TEST_SOURCES" FORCE)
+
